+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Change `ZestClientElementScrollTo` to scroll with `nearest` vertical alignment, to ensure the element is kept visible.
 
 ## [0.28.0] - 2025-05-15
 ### Changed